20210107 ss

+++ mcr-test/src/main/java/external/instrumented/java/util/Random.java
@@ -476,7 +476,7 @@ class Random implements java.io.Serializable {
         ObjectInputStream.GetField fields = s.readFields();
         long seedVal;
 
-        seedVal = (long) fields.get("seed", -1L);
+        seedVal = fields.get("seed", -1L);
         if (seedVal < 0)
           throw new java.io.StreamCorruptedException(
                               "Random: invalid seed");
